# feat(performance): improve core undenting performance by 20-30x